requirements.txt
tensorflow==2.10.1
numpy==1.23.5
scikit-learn==1.3.2
matplotlib==3.7.3
pandas==2.2.2
Pillow==10.3.0
tqdm==4.66.4
nltk==3.8.1
kaggle==1.6.12
flask==3.0.3
waitress==3.0.0
protobuf<4